Proposing rule for referencing class methods in other class methods.
Using the example of a "service object" and its standard `.call` class
method that instantiates the class it's a member of, this proposer
suggests that class methods simply call other class methods, rather than
regurgitating the class name or even the `self` keyword. Many people,
including this proposer, tend to forget that `new` will (generally) Do
The Right Thing in such situations without decoration *because* the
method in `BasicObject` is overridden appropriately in `Class` and
elsewhere.